# openapi.yaml
openapi: 3.1.0
info:
  title: FakeStore API
  version: 1.0.0
  description: FakeStore backend API with layered architecture, JWT auth, and carts/ratings.
paths:
...
  /api/products/:
    get:
      operationId: products_list
      description: Supports pagination via ?page and ?limit. Cached results may be
        served.
      summary: List products
      parameters:
      - in: query
        name: category
        schema:
          type: string
        description: Filter by category name
      tags:
      - Catalog
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductRead'
          description: ''
    post:
      operationId: products_create
      summary: Create product
      tags:
      - Catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductRead'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/products/{product_id}/:
    get:
      operationId: products_retrieve
      summary: Get product
      parameters:
      - in: path
        name: product_id
        schema:
          type: integer
        required: true
      tags:
      - Catalog
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductRead'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
    put:
      operationId: products_update
      summary: Replace product
      parameters:
      - in: path
        name: product_id
        schema:
          type: integer
        required: true
      tags:
      - Catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductRead'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
    patch:
      operationId: products_partial_update
      summary: Update product
      parameters:
      - in: path
        name: product_id
        schema:
          type: integer
        required: true
      tags:
      - Catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProductWriteRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductRead'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
    delete:
      operationId: products_destroy
      summary: Delete product
      parameters:
      - in: path
        name: product_id
        schema:
          type: integer
        required: true
      tags:
      - Catalog
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
      required:
      - id
      - name
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
      required:
      - code
      - message
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      required:
      - error
...
    PaginatedProductRead:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/ProductRead'
      required:
      - count
      - next
      - previous
      - results
...
    PatchedProductWriteRequest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          minLength: 1
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        description:
          type: string
          minLength: 1
        image:
          type: string
        rate:
          type: number
          format: double
        count:
          type: integer
        categories:
          type: array
          items:
            type: integer
...
    ProductRead:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        price:
          type: string
        description:
          type: string
        image:
          type: string
        rate:
          type: string
        count:
          type: integer
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
      required:
      - categories
      - count
      - description
      - id
      - image
      - price
      - rate
      - title
    ProductWriteRequest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          minLength: 1
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        description:
          type: string
          minLength: 1
        image:
          type: string
        rate:
          type: number
          format: double
        count:
          type: integer
        categories:
          type: array
          items:
            type: integer
      required:
      - description
      - id
      - image
      - price
      - title
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT